4c7cc5d COMP: Fix some implicit cast warnings

+++ Utilities/gdcm/Source/Common/gdcmMD5.cxx
@@ -45,7 +45,7 @@ MD5::~MD5()
   delete Internals;
 }
 
-bool MD5::Compute(const char *buffer, unsigned long buf_len, char digest_str[33])
+bool MD5::Compute(const char *buffer, size_t buf_len, char digest_str[33])
 {
   if( !buffer || !buf_len )
     {
